+++ secretspec/src/provider/mod.rs
@@ -159,6 +159,7 @@ pub mod gcsm;
 pub mod keyring;
 pub mod lastpass;
 pub mod onepassword;
+pub mod onepassword_env;
 pub mod pass;
 pub mod protonpass;
 #[cfg(feature = "vault")]